refactor(variable): move update confirmation to service layer #179
