+++ c8volt/process/client_test.go
@@ -13,7 +13,6 @@ import (
 	"time"
 
 	options "github.com/grafvonb/c8volt/c8volt/foptions"
-	"github.com/grafvonb/c8volt/config"
 	d "github.com/grafvonb/c8volt/internal/domain"
 	"github.com/grafvonb/c8volt/internal/services"
 	pdsvc "github.com/grafvonb/c8volt/internal/services/processdefinition"
@@ -437,7 +436,7 @@ func TestClient_EnrichProcessInstancesWithVariables_SortsVariablesAndPreservesJS
 	}, got.Items[0].Variables)
 }
 
-func TestVariableConfirmation_ConfirmsRequestedVariablesWithNormalizedJSON(t *testing.T) {
+func TestUpdateProcessInstanceVariablesMapsConfirmedServiceResponse(t *testing.T) {
 	t.Parallel()
 
 	ctx := context.Background()
@@ -450,16 +449,6 @@ func TestVariableConfirmation_ConfirmsRequestedVariablesWithNormalizedJSON(t *te
 			}, variables)
 			assert.True(t, services.ApplyCallOptions(opts).Verbose)
 			return d.ProcessInstanceVariableUpdateResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-		},
-		searchProcessInstanceVariables: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceVariable, error) {
-			require.Equal(t, "123", key)
-			assert.True(t, services.ApplyCallOptions(opts).Verbose)
-			return []d.ProcessInstanceVariable{
-				{Name: "foo", Value: `"bar"`, ProcessInstanceKey: "123", ScopeKey: "123"},
-				{Name: "nested", Value: `{"count":2}`, ProcessInstanceKey: "123", ScopeKey: "123"},
-				{Name: "unrelated", Value: `"ignored"`, ProcessInstanceKey: "123", ScopeKey: "123"},
-				{Name: "foo", Value: `"wrong-scope"`, ProcessInstanceKey: "123", ScopeKey: "element-1"},
-			}, nil
 		},
 	}
 
@@ -485,48 +474,6 @@ func TestVariableConfirmation_ConfirmsRequestedVariablesWithNormalizedJSON(t *te
 			"nested": map[string]any{"count": float64(2)},
 		},
 	}, got)
-}
-
-func TestVariableConfirmation_WaitsUntilRequestedVariablesAreVisible(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	var lookups atomic.Int32
-	piAPI := stubProcessInstanceAPI{
-		updateProcessInstanceVariables: func(_ context.Context, key string, variables map[string]any, opts ...services.CallOption) (d.ProcessInstanceVariableUpdateResponse, error) {
-			require.Equal(t, "123", key)
-			require.Equal(t, map[string]any{"hasIncident": true}, variables)
-			return d.ProcessInstanceVariableUpdateResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-		},
-		searchProcessInstanceVariables: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceVariable, error) {
-			require.Equal(t, "123", key)
-			if lookups.Add(1) == 1 {
-				return []d.ProcessInstanceVariable{
-					{Name: "hasIncident", Value: `false`, ProcessInstanceKey: "123", ScopeKey: "123"},
-				}, nil
-			}
-			return []d.ProcessInstanceVariable{
-				{Name: "hasIncident", Value: `true`, ProcessInstanceKey: "123", ScopeKey: "123"},
-			}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, piAPI, slog.Default())
-	got, err := cli.UpdateProcessInstanceVariables(ctx, ProcessInstanceVariableUpdateRequest{
-		Key:       "123",
-		Variables: map[string]any{"hasIncident": true},
-	}, options.WithBackoff(config.BackoffConfig{
-		Strategy:     config.BackoffFixed,
-		InitialDelay: time.Millisecond,
-		MaxDelay:     time.Millisecond,
-		MaxRetries:   3,
-		Timeout:      time.Second,
-	}))
-
-	require.NoError(t, err)
-	require.Equal(t, int32(2), lookups.Load())
-	require.Equal(t, ProcessInstanceVariableUpdateStatusConfirmed, got.Status)
-	require.True(t, got.OK())
 }
 
 func TestUpdateProcessInstanceVariablesMultipleKeysRespectWorkersAndFailFastOptions(t *testing.T) {
@@ -646,11 +593,7 @@ func TestUpdateProcessInstanceVariablesConfirmationTimeoutReportsPerKeyFailure(t
 		updateProcessInstanceVariables: func(_ context.Context, key string, variables map[string]any, opts ...services.CallOption) (d.ProcessInstanceVariableUpdateResponse, error) {
 			require.Equal(t, "123", key)
 			require.Equal(t, map[string]any{"foo": "bar"}, variables)
-			return d.ProcessInstanceVariableUpdateResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-		},
-		searchProcessInstanceVariables: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceVariable, error) {
-			require.Equal(t, "123", key)
-			return nil, context.DeadlineExceeded
+			return d.ProcessInstanceVariableUpdateResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, context.DeadlineExceeded
 		},
 	}
