+++ internal/services/processinstance/v88/convert.go
@@ -4,8 +4,6 @@
 package v88
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
 	"time"
 
@@ -46,57 +44,6 @@ func fromIncidentResult(r camundav88.IncidentResult) d.ProcessInstanceIncidentDe
 		ProcessDefinitionKey:   r.ProcessDefinitionKey,
 		ProcessDefinitionId:    r.ProcessDefinitionId,
 	}
-}
-
-type variableSearchQueryResult struct {
-	Items []variableSearchResult             `json:"items"`
-	Page  camundav88.SearchQueryPageResponse `json:"page"`
-}
-
-type variableSearchResult struct {
-	Name               string `json:"name"`
-	Value              string `json:"value"`
-	VariableKey        string `json:"variableKey"`
-	ProcessInstanceKey string `json:"processInstanceKey"`
-	ScopeKey           string `json:"scopeKey"`
-	TenantId           string `json:"tenantId"`
-	IsTruncated        *bool  `json:"isTruncated,omitempty"`
-	Truncated          *bool  `json:"truncated,omitempty"`
-}
-
-func fromVariableSearchResult(r variableSearchResult) d.ProcessInstanceVariable {
-	return d.ProcessInstanceVariable{
-		Name:               r.Name,
-		Value:              r.Value,
-		VariableKey:        r.VariableKey,
-		ProcessInstanceKey: r.ProcessInstanceKey,
-		ScopeKey:           r.ScopeKey,
-		TenantId:           r.TenantId,
-		APITruncated:       variableAPITruncated(r),
-	}
-}
-
-// decodeSearchVariablesResponse reads raw JSON because the generated v8.8 model drops value and truncation fields.
-func decodeSearchVariablesResponse(body []byte, page *camundav88.VariableSearchQueryResult) (variableSearchQueryResult, error) {
-	if len(bytes.TrimSpace(body)) == 0 {
-		return variableSearchQueryResult{}, d.ErrMalformedResponse
-	}
-	var result variableSearchQueryResult
-	if err := json.Unmarshal(body, &result); err != nil {
-		return variableSearchQueryResult{}, err
-	}
-	if page != nil {
-		result.Page = page.Page
-	}
-	return result, nil
-}
-
-// variableAPITruncated accepts both observed field names so older and newer API payloads keep the same display contract.
-func variableAPITruncated(r variableSearchResult) bool {
-	if r.IsTruncated != nil {
-		return *r.IsTruncated
-	}
-	return toolx.Deref(r.Truncated, false)
 }
 
 func formatTime(t time.Time) string {
