+++ internal/services/processinstance/v88/service_test.go
@@ -1127,7 +1127,7 @@ func TestElementInstanceVariablesUpdate_UsesProcessInstanceKeyAsElementInstanceK
 		cancelProcessInstanceWithResponse: unexpectedCancelProcessInstance(t),
 	}, newStrictOperateClient(t))
 
-	got, err := svc.UpdateProcessInstanceVariables(ctx, "123", map[string]any{"foo": "bar"})
+	got, err := svc.UpdateProcessInstanceVariables(ctx, "123", map[string]any{"foo": "bar"}, services.WithNoWait())
 
 	require.NoError(t, err)
 	assert.Equal(t, d.ProcessInstanceVariableUpdateResponse{
