+++ internal/services/processinstance/v89/convert.go
@@ -48,57 +48,6 @@ func fromIncidentResult(r camundav89.IncidentResult) d.ProcessInstanceIncidentDe
 	}
 }
 
-type variableSearchQueryResult struct {
-	Items []variableSearchResult             `json:"items"`
-	Page  camundav89.SearchQueryPageResponse `json:"page"`
-}
-
-type variableSearchResult struct {
-	Name               string `json:"name"`
-	Value              string `json:"value"`
-	VariableKey        string `json:"variableKey"`
-	ProcessInstanceKey string `json:"processInstanceKey"`
-	ScopeKey           string `json:"scopeKey"`
-	TenantId           string `json:"tenantId"`
-	IsTruncated        *bool  `json:"isTruncated,omitempty"`
-	Truncated          *bool  `json:"truncated,omitempty"`
-}
-
-func fromVariableSearchResult(r variableSearchResult) d.ProcessInstanceVariable {
-	return d.ProcessInstanceVariable{
-		Name:               r.Name,
-		Value:              r.Value,
-		VariableKey:        r.VariableKey,
-		ProcessInstanceKey: r.ProcessInstanceKey,
-		ScopeKey:           r.ScopeKey,
-		TenantId:           r.TenantId,
-		APITruncated:       variableAPITruncated(r),
-	}
-}
-
-// decodeSearchVariablesResponse reads raw JSON because the generated v8.9 model drops value and truncation fields.
-func decodeSearchVariablesResponse(body []byte, page *camundav89.VariableSearchQueryResult) (variableSearchQueryResult, error) {
-	if len(bytesTrimSpace(body)) == 0 {
-		return variableSearchQueryResult{}, d.ErrMalformedResponse
-	}
-	var result variableSearchQueryResult
-	if err := json.Unmarshal(body, &result); err != nil {
-		return variableSearchQueryResult{}, err
-	}
-	if page != nil {
-		result.Page = page.Page
-	}
-	return result, nil
-}
-
-// variableAPITruncated accepts both observed field names so older and newer API payloads keep the same display contract.
-func variableAPITruncated(r variableSearchResult) bool {
-	if r.IsTruncated != nil {
-		return *r.IsTruncated
-	}
-	return toolx.Deref(r.Truncated, false)
-}
-
 func formatTime(t time.Time) string {
 	if t.IsZero() {
 		return ""
