+++ internal/services/variable/v87/contract.go
@@ -10,7 +10,6 @@ import (
 )
 
 type GenVariableClientOperate interface {
-	GetVariableByKeyWithResponse(ctx context.Context, key int64, reqEditors ...operatev87.RequestEditorFn) (*operatev87.GetVariableByKeyResponse, error)
 	SearchVariablesForProcessInstancesWithResponse(ctx context.Context, body operatev87.SearchVariablesForProcessInstancesJSONRequestBody, reqEditors ...operatev87.RequestEditorFn) (*operatev87.SearchVariablesForProcessInstancesResponse, error)
 }
