+++ internal/services/variable/v88/contract.go
@@ -10,8 +10,8 @@ import (
 )
 
 type GenVariableClientCamunda interface {
-	GetVariableWithResponse(ctx context.Context, variableKey string, reqEditors ...camundav88.RequestEditorFn) (*camundav88.GetVariableResponse, error)
 	SearchVariablesWithResponse(ctx context.Context, params *camundav88.SearchVariablesParams, body camundav88.SearchVariablesJSONRequestBody, reqEditors ...camundav88.RequestEditorFn) (*camundav88.SearchVariablesResponse, error)
+	CreateElementInstanceVariablesWithResponse(ctx context.Context, elementInstanceKey camundav88.ElementInstanceKey, body camundav88.CreateElementInstanceVariablesJSONRequestBody, reqEditors ...camundav88.RequestEditorFn) (*camundav88.CreateElementInstanceVariablesResponse, error)
 }
 
 var _ GenVariableClientCamunda = (*camundav88.ClientWithResponses)(nil)
